`spack python`: use `Executable` instead of `InteractiveConsole`
Running a `spack-python` script like this:

```python

import spack
import multiprocessing

def echo(args):
    print(args)

if __name__ == "__main__":
    pool = multiprocessing.Pool(2)
    pool.map(echo, range(10))
```

will fail in `develop` with an error like this:

```console
_pickle.PicklingError: Can't pickle <function echo at 0x104865820>: attribute lookup echo on __main__ failed
```

Python expects to be able to look up the method `echo` in `sys.path["__main__"]` in
subprocesses spawned by `multiprocessing`, but because we use `InteractiveConsole` to
run `spack python`, the executed file isn't considered to be the `__main__` module, and
lookups in subprocesses fail. We tried to fake this by setting `__name__` to `__main__`
in the `spack python` command, but that doesn't fix the fact that no `__main__` module
exists.

Another annoyance with `InteractiveConsole` is that `__file__` is not defined in the
main script scope, so you can't use it in your scripts.

We can fix this by using `spack.util.executable.Executable` to launch `spack python`
instead of using `InteractiveConsole`. This makes `spack python` behave more like
normal python. After this change:

```console
> ./test.py
0
1
2
3
4
5
6
7
8
9
```

- [x] Use `Executable` to launch non-interactive `spack python` invocations
- [x] Only use `InteractiveConsole` for interactive `spack python`